update_ProductOptionClear (#157)
* Login update

Login update

* Update Login

Update Login

* update-admin-frontend

update-admin-frontend

* update_Coupon

update_Coupon

* update_AdminCopon

update_AdminCopon

* update_AdminCouponClear

update_AdminCouponClear

* update_AdminProduct_Clear

update_AdminProduct_Clear

* update_AdminShowPage

update_AdminShowPage

* update_AdminProductOption

추후 재 업데이트 예정

* update_ProductOptionClear

update_ProductOptionClear

+++ 01_Backend/backend/src/main/java/org/example/backend/service/mypage/MyWishlistService.java
@@ -45,7 +45,7 @@ public class MyWishlistService {
 
     // pdId를 사용한 소프트 삭제 함수
     @Transactional
-    public void removeWishlistItem(Integer pdId, String userId) {
+    public void removeWishlistItem(Long pdId, String userId) {
         myWishlistRepository.deleteById(new PdIdUserIdPk(pdId, userId));
     }
     }
